requirements.txt
streamlit>=1.30.0
finance-datareader>=0.9.0
yfinance>=0.2.0
pykrx>=1.0.0
anthropic>=0.40.0
openai>=1.0.0
ta>=0.11.0
plotly>=5.18.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
matplotlib
deprecated
multipledispatch
lxml
html5lib
